Refactor envelope tests into focused modules
